github-318: move performance testing classes to a separate tests package
