+++ src/Truncation/StringsStrategy.php
@@ -32,7 +32,8 @@ class StringsStrategy extends AbstractStrategy
         }
         
         if ($modified) {
-            $payload = new EncodedPayload($data);
+            $payloadClass = get_class($payload);
+            $payload = new $payloadClass($data);
             $payload->encode();
         }
