+++ tests/BaseRollbarTest.php
@@ -7,7 +7,6 @@ abstract class BaseRollbarTest extends \PHPUnit_Framework_TestCase
     
     public function tearDown()
     {
-        \Rollbar\Payload\EncodedPayload::resetEncodingCount();
         Rollbar::destroy();
     }
